+++ src/main/java/domain/board/pathfinder/PathFinderFactory.java
@@ -1,0 +1,6 @@
+package domain.board.pathfinder;
+
+public interface PathFinderFactory {
+
+    PathFinder createPathFinder();
+}
