+++ src/test/java/fixture/BoardFixture.java
@@ -1,18 +1,18 @@
 package fixture;
 
 import domain.board.Board;
-import domain.board.PathFinder;
-import domain.board.PathFinderFactory;
+import domain.board.pathfinder.DefaultPathFinderFactory;
+import domain.board.pathfinder.PathFinder;
 import domain.piece.Piece;
 import domain.point.Point;
 import java.util.Map;
 
 public class BoardFixture {
 
-    private static final PathFinderFactory pathFinderFactory = new PathFinderFactory();
+    private static final DefaultPathFinderFactory DEFAULT_PATH_FINDER_FACTORY = DefaultPathFinderFactory.getInstance();
 
     public static PathFinder createDefaultPathFinder() {
-        return pathFinderFactory.createDefaultPathFinder();
+        return DEFAULT_PATH_FINDER_FACTORY.createPathFinder();
     }
 
     public static Board createTestBoard(Map<Point, Piece> pieceByPoint) {
